\name{SEAHORS.MICRO}
\alias{SEAHORS.MICRO}

\title{SEAHORS.MICRO}

\description{This application is a R shiny application dedicated to the intra-site spatial analysis of piece-plotted archaeological remains, which were recorded using independant subdivision (splits) with Cartesian coordinates. This application making the two and three-dimensional spatial exploration of archaeological data as user-friendly as possible.}

\usage{
SEAHORS.MICRO()
}

\details{This is a launcher function for the `SEAHORS.MICRO` application.}

\value{Launch the `SEAHORS.MICRO` Shiny application.}

\author{Aurelien Royer <aurelien.royer at u-bourgogne.fr>}

\seealso{
	\code{shiny}
	\code{plotly}
	\code{ggplot2}
	\code{SEAHORS}
}

\references{

}

\examples{
	if(interactive()){SEAHORS.MICRO()}
}
